requirements.txt
asgiref==3.11.0
dj-rest-auth==7.0.1
Django==4.2.26
django-cors-headers==4.9.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
MarkupSafe==3.0.3
mongoengine==0.29.1
PyJWT==2.10.1
pymongo==3.11.4
python-decouple==3.8
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.4
Werkzeug==3.1.4
gunicorn==21.2.0
whitenoise==6.7.0
